+++ src/main/java/org/informatics/service/impl/StoreServiceImpl.java
@@ -2,6 +2,7 @@ package org.informatics.service.impl;
 
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 
 import org.informatics.entity.Cashier;
 import org.informatics.entity.Customer;
@@ -51,7 +52,7 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public double turnover() {
+    public BigDecimal turnover() {
         return store.turnover();
     }
 }
